+++ packages/partitura_cli/test/gp_fixtures_test.dart
@@ -1,6 +1,5 @@
 import 'dart:io';
 
-import 'package:partitura_cli/src/gp_container.dart';
 import 'package:partitura_core/partitura_core.dart';
 import 'package:test/test.dart';